Fix compilation for mingw64. 	* coffcode.h (coff_slurp_symbol_table): Add intptr_t intermediate 	typecast to avoid warning. 	* elf32-rx.c: Add "bfd_stdint.h" include required for int32_t type 	usage. 	* elfxx-ia64.c (elfNN_ia64_relax_br): Use intptr_t typeacast instead 	of long for pointer to avoid warning. 	(elfNN_ia64_relax_brl): Idem. 	(elfNN_ia64_install_value): Idem. 	* vms-alpha.c (_bfd_vms_slurp_etir): Idem.

+++ bfd/coffcode.h
@@ -4852,7 +4852,7 @@ coff_slurp_symbol_table (bfd * abfd)
 		 to the symbol instead of the index.  FIXME: This
 		 should use a union.  */
 	      src->u.syment.n_value =
-		(long) (native_symbols + src->u.syment.n_value);
+		(long) (intptr_t) (native_symbols + src->u.syment.n_value);
 	      dst->symbol.value = src->u.syment.n_value;
 	      src->fix_value = 1;
 	      break;
